fix: Move Advent calendar date logic to client-side for static export compatibility
